Exercise 1 and 2 ok
