+++ exercise/28_2/index.js
@@ -1,12 +1,14 @@
 const express = require('express');
 const bodyParser = require('body-parser');
+const productsController = require('./controllers/productController');
+
 
 const app = express();
 app.use(bodyParser.json());
 
 app.use(bodyParser.urlencoded({ extended: false }));
 
-app.use('/products/', require('./controllers/productController'));
+app.use('/products', productsController);
 
 app.listen(3000, () => {
   console.log("App listening on port 3000!");
