+++ exercise/28_2/products.json
@@ -1,0 +1,1 @@
+[{"id":1,"name":"Cerveja Skol","brand":"Ambev"},{"id":2,"name":"Monitor AGON","brand":"AOC"},{"id":3,"name":"MacBook Air","brand":"Apple"}]
